% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_api_units.R
\name{account_api_units}
\alias{account_api_units}
\title{View SEMRush API Units balance for a subscriber account.}
\usage{
account_api_units(key, character_format = FALSE)
}
\arguments{
\item{key}{\emph{string}. API key assigned to a user after subscribing to SEMrush. The key is available from the account profile page.}

\item{character_format}{\emph{logical}. If FALSE (default), returns the unit balance as a numeric. If TRUE, value returned is a character string formatted with a "," every three digits.}
}
\description{
Returns the SEMRush account API Units balance. API Units are consumed when requests to the SEMRush database are made through the API.
See the SEMRush API website \href{https://www.semrush.com/api-analytics/}{(https://www.semrush.com/api-analytics/)} for more information.
}
\examples{
##define object with your SEMRush account API key.
\dontrun{
api_key <- ""

## without formatting
account_api_units(api_key)
10000

## with formatting
account_api_units(api_key, character_format=TRUE)
"10,000"}
}
