% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advertiser_reports.R
\name{advertising_reports}
\alias{advertising_reports}
\title{Request SEMRush Domain Reports}
\usage{
advertising_reports(
  type,
  key,
  domain,
  advertiser_domain,
  publisher_domain,
  display_limit = 5,
  display_offset,
  device_type,
  export_columns,
  return_url = FALSE
)
}
\arguments{
\item{type}{string. The report type to be generated. Available report types are "publisher_text_ads", "publisher_advertisers", "advertiser_publishers", "advertiser_text_ads", "advertiser_landings", "advertiser_publisher_text_ads", "advertiser_rank", or "publisher_rank".}

\item{key}{string. An identification key assigned to a user after subscribing to SEMrush. The key is available on the Profile page.}

\item{domain}{string. A unique name of a website you would like to investigate (example.com).}

\item{advertiser_domain}{string. A unique name of a website you would like to investigate (example.com). Used for 'advertiser_publisher_text_ads' report only.}

\item{publisher_domain}{string. A unique name of a website you would like to investigate (example.com). Used for 'advertiser_publisher_text_ads' report only.}

\item{display_limit}{integer. The number of results returned to a request. If this parameter is set to 0, up to 10,000 lines will be returned. Default is 5 lines.}

\item{display_offset}{integer. This parameter allows you to skip a specified number of results before sending a report. Please keep in mind that if you use the option display_offset, display_limit, the value should be increased by value display_offset.}

\item{device_type}{string. This parameter shows the type of device by which statistics have been collected-a PC, tablet, or smartphone. Only used for some report types. Options: "all", "desktop", "smartphone_apple", "smartphone_android", "tablet_apple", or "tablet_android".}

\item{export_columns}{\emph{vector}. A vector of character strings specifying the variables to be included in the report, which vary according to the report type (see 'type' argument). If this parameter is not specified, default columns will be sent.}

\item{return_url}{logical. If TRUE, prints the request URL used to generate the report. Default value is FALSE.}
}
\value{
A data table (tibble) with columns for each requested variable.
}
\description{
This function creates a request to the SEMRush SEO database for users with a subscription and sufficient API units.
}
\examples{
## Enter your SEMRush account API key
\dontrun{
key <- "" #enter your SEMRush account API key.

##Get 'backlinks_overview' report
advertising_reports(
type = "advertiser_landings",
domain = "cran.r-project.org",
key = key,
device_type = "all",
return_url = FALSE
)

}
}
