% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_reports.R
\name{domain_reports}
\alias{domain_reports}
\title{Request SEMRush Domain Reports}
\usage{
domain_reports(
  type,
  key,
  domain,
  domains,
  database = "us",
  display_limit = 5,
  display_offset,
  display_date,
  export_columns,
  display_daily,
  return_url = FALSE
)
}
\arguments{
\item{type}{string. The report type to be generated. Available report types are 'domain_organic', 'domain_adwords', 'domain_adwords_unique', 'domain_adwords_adwords', 'domain_adwords_historical', 'domain_domains', 'domain_shopping', 'domain_shopping_unique', 'domain_shopping_shopping', 'domain_organic_unique', and 'domain_organic_subdomains'.}

\item{key}{string. An identification key assigned to a user after subscribing to SEMrush. The key is available on the Profile page.}

\item{domain}{string. A unique name of a website you would like to investigate (example.com).}

\item{domains}{vector. A vector of strings for each domain name that contains domains in a specified format, separated by "|". This format requires a domain to consist of <sign>|<type>|<domain>, where the <sign> means possible operations: "+", "-", "*", "/"; <type> has to take the value of "or" for organic keywords or "ad" for paid keywords; and( <domain> indicates a domain.}

\item{database}{string. A regional database. If this parameter is not specified (i.e., ""), your request will be sent to all regional databases. Default is the United States.}

\item{display_limit}{integer. The number of results returned to a request. If this parameter is set to 0, up to 10,000 lines will be returned. Default is 5 lines.}

\item{display_offset}{integer. This parameter allows you to skip a specified number of results before sending a report. Please keep in mind that if you use the option display_offset, display_limit, the value should be increased by value display_offset.}

\item{display_date}{string. Date in format "YYYYMM15". A date on which a report will be shown. You can roll back to the past or choose an actual date.}

\item{export_columns}{\emph{vector}. A vector of character strings specifying the variables to be included in the report, which vary according to the report type (see 'type' argument). If this parameter is not specified, default columns will be sent.}

\item{display_daily}{integer. Only used for some report types. This parameter allows you to get daily updates on position changes that occurred in the last 30 days or more. If the parameter is not specified, your report will show monthly results for the current month and for previous months.}

\item{return_url}{logical. If TRUE, prints the request URL used to generate the report. Default value is FALSE.}
}
\value{
A data table (tibble) with columns for each requested variable.
}
\description{
This function creates a request to the SEMRush SEO database for users with a subscription and sufficient API units.
}
\examples{
## Enter your SEMRush account API key
key <- ""

## Request example for domain_domains

#define 'domains' string, using the format specificed from https://www.semrush.com/api-analytics/#domain_domains.
domains <- "*|or|nike.com|*|or|adidas.com|*|or|reebok.com"

#Use function to request report
report <-
  domain_reports(
    type = "domain_domains",
    key = key,
    domains = domains,
    database = "us"
)

print(report)
# A tibble: 5 x 12
Keyword nike.com adidas.com reebok.com Domain4.Pos Domain5.Pos Number.of.Resul…   CPC Search.Volume Keyword.Difficu…
<chr>      <int>      <int>      <int>       <int>       <int>            <dbl>   <dbl>         <int>            <dbl>
1 shoes         62         35         36           0           0       4750000000  1.08       1500000             90.0
2 basket…        3         14         13           0           0        339000000  0.64        368000             88.9
3 man           40         58         64           0           0      25270000000  0.46        301000             87.8
4 shoes …       24         57         55           0           0       6080000000  0.88        301000             91.5
5 shoes …       63         38         29           0           0       4710000000  1.01        301000             91.0
}
