% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_reports.R
\name{overview_reports}
\alias{overview_reports}
\title{Request SEMRush Overview Reports}
\usage{
overview_reports(
  type,
  key,
  domain,
  database,
  display_limit = 5,
  display_offset,
  display_date,
  export_columns,
  display_daily,
  return_url = FALSE
)
}
\arguments{
\item{type}{\emph{string}. The report type to be generated. Available report types are 'domain_ranks','domain_rank','domain_rank_history','rank_difference', or 'rank'.}

\item{key}{\emph{string}. An identification key assigned to a user after subscribing to SEMrush. The key is available on the Profile page.}

\item{domain}{\emph{string}. A unique name of a website you would like to investigate (example.com).}

\item{database}{\emph{string}. A regional database. If this parameter is not specified (i.e., ""), your request will be sent to all regional databases.}

\item{display_limit}{\emph{integer}. The number of results returned to a request. If this parameter is set to 0, up to 10,000 lines will be returned. Default is 5 lines.}

\item{display_offset}{\emph{integer}. This parameter allows you to skip a specified number of results before sending a report. Please keep in mind that if you use the option display_offset, display_limit, the value should be increased by value display_offset.}

\item{display_date}{\emph{string}. Date in format "YYYYMM15". A date on which a report will be shown. You can roll back to the past or choose an actual date.}

\item{export_columns}{\emph{vector}. A vector of character strings specifying the variables to be included in the report, which vary according to the report type (see 'type' argument). If this parameter is not specified, default columns will be sent.}

\item{display_daily}{\emph{integer}. Only used for some report types. This parameter allows you to get daily updates on position changes that occurred in the last 30 days or more. If the parameter is not specified, your report will show monthly results for the current month and for previous months.}

\item{return_url}{logical. If TRUE, prints the request URL used to generate the report. Default value is FALSE.}
}
\value{
A data table (tibble) with columns for each requested variable.
}
\description{
This function creates a request to the SEMRush SEO database for users with a subscription and sufficient API units.
See the SEMRush API website \href{https://www.semrush.com/api-analytics/}{(https://www.semrush.com/api-analytics/)} for additional information, including a list of codes for the regional databases and export variables.
}
\examples{
#Enter your SEMRush account API key
key <- ""

## Get 'domain_ranks' report for a single domain
report <- overview_reports(
  type = "domain_ranks",
  key = key,
  domain = "cran.r-project.org",
  display_limit = 5
)

print(report)
# A tibble: 5 x 11
Database   Domain  Rank Organic.Keywords Organic.Traffic Organic.Cost Adwords.Keywords Adwords.Traffic Adwords.Cost
<chr>      <chr>  <int>            <int>           <int>        <int>            <int>           <int>        <int>
1 ph       r-pro…   336             8980          504281        54705                0               0            0
2 ge       r-pro…   865              593            8319          799                0               0            0
3 et       r-pro…   931             1196            2496          141                0               0            0
4 kh       r-pro…   996              648            5122          446                0               0            0
5 is       r-pro…  1012              656            2202            1                0               0            0
… with 2 more variables: PLA.uniques <int>, PLA.keywords <int>

## Get 'rank' report
report <- overview_reports(
  type = "rank",
  key = api_key,
  domain = "cran.r-project.org",
  database = "us", #United States
  display_limit = 5
)

print(report)
# A tibble: 5 x 8
  Domain         Rank Organic.Keywords Organic.Traffic Organic.Cost Adwords.Keywords Adwords.Traffic Adwords.Cost
  <chr>         <int>            <int>           <dbl>        <dbl>            <int>           <int>        <int>
1 wikipedia.org     1        105359860      3038929100   4065812051                1               1            0
2 youtube.com       2         95798001      1156183092    801586911             2629          409509      1193302
3 amazon.com        3         93291441       998829876    805881241            77380         3103766      9102331
4 google.com        4        484166619       662298708   2701485008            20414         4318917     63404537
5 facebook.com      5         97924346       551902310    560771216             7729          647205      2725034
}
