% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_keyword_reports.R
\name{sem_keyword_reports}
\alias{sem_keyword_reports}
\title{Request SEMRush Keyword Reports}
\usage{
sem_keyword_reports(
  type,
  key,
  phrase,
  database,
  display_limit = 5,
  display_offset,
  display_date,
  export_columns,
  return_url = FALSE
)
}
\arguments{
\item{type}{\emph{string}. REQUIRED. The report type to be generated. Available report types are 'phrase_all', 'phrase_this','phrase_these','phrase_organic','phrase_adwords',
'phrase_related',
'phrase_adwords_historical',
'phrase_fullsearch',
'phrase_questions', or
'phrase_kdi'}

\item{key}{\emph{string}. REQUIRED. An identification key assigned to a user after subscribing to SEMrush. The key is available on the Profile page.}

\item{phrase}{\emph{string, vector}. REQUIRED. A keyword or keyword expression you'd like to investigate. Also, a vector of character strings.}

\item{database}{string. A regional database. If this parameter is not specified (i.e., ""), your request will be sent to all regional databases, if possible. Default is the United States.}

\item{display_limit}{\emph{integer}. The number of results returned to a request. If this parameter is set to 0, up to 10,000 lines will be returned. Default is 5 lines.}

\item{display_offset}{\emph{integer}. This parameter allows you to skip a specified number of results before sending a report. Please keep in mind that if you use the option display_offset, display_limit, the value should be increased by value display_offset.}

\item{display_date}{\emph{string}. Date in format "YYYYMM15". A date on which a report will be shown. You can roll back to the past or choose an actual date.}

\item{export_columns}{\emph{vector}. A vector of character strings specifying the variables to be included in the report, which vary according to the report type (see 'type' argument). If this parameter is not specified, default columns will be sent.}

\item{return_url}{logical. If TRUE, prints the request URL used to generate the report. Default value is FALSE.}
}
\value{
A data table (tibble) with columns for each requested variable.
}
\description{
This function creates a request to the SEMRush SEO database for users with a subscription and sufficient API units.
See the SEMRush API website \href{https://www.semrush.com/api-analytics/}{(https://www.semrush.com/api-analytics/)} for additional information, including a list of codes for the regional databases and export variables.
}
\examples{
#Enter your SEMRush account API key
key <- ""

## Generate 'phrase_organic' report

report <- sem_keyword_reports(
    type = "phrase_organic",
    key = key,
    phrase = "r software",
    database = "us",
    display_limit=10
)

print(report)
A tibble: 10 x 2
   Domain                        Url
   <chr>                         <chr>
 1 r-project.org                 https://www.r-project.org/
 2 wikipedia.org                 https://en.wikipedia.org/wiki/R_(programming_language)
 3 rstudio.com                   https://rstudio.com/
 4 umich.edu                     https://www.icpsr.umich.edu/icpsrweb/content/shared/ICPSR/faqs/what-is-r.html
 5 epa.gov                       https://archive.epa.gov/nheerl/arm/web/pdf/irss_2.6.pdf
 6 psu.edu                       https://online.stat.psu.edu/statprogram/tutorials/statistical-software/r
 7 datamentor.io                 https://www.datamentor.io/r-programming/
 8 statmethods.net               https://www.statmethods.net/r-tutorial/index.html
 9 utoledo.edu                   https://libguides.utoledo.edu/stats-software/R
10 predictiveanalyticstoday.com https://www.predictiveanalyticstoday.com/r-software-environment/

## Generate 'phrase_these' report

report <- sem_keyword_reports(
    type = "phrase_these",
    key = key,
    phrase = c("statistical programming","r software"),
    database = "us",
    display_limit=10
)

print(report)
# A tibble: 2 x 6
Keyword                 Search.Volume CPC   Competition Number.of.Results Trends
<chr>                       <int>     <dbl>   <dbl>     <dbl>             <chr>
1 r software                 4400     2.68    0         4190000000      0.67,0.67,0.55,0.67,0.82,0.82,0.82,0.67,0.55,1.0…
2 statistical programm…      1000     3.91    0.14      394000000       0.81,0.63,0.63,0.55,0.63,0.81,0.81,0.63,0.55,1.0…
}
