% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semrush_reports.R
\name{semrush_reports}
\alias{semrush_reports}
\title{Wrapper function for all SEMRush report types.}
\usage{
semrush_reports(report, ...)
}
\arguments{
\item{report}{\emph{string}. The report type on which to pass all additional arguments. Choices are 'advertising', 'backlinks', 'domain', 'keywords', or 'overview'.}

\item{`...`}{Additional arguments to pass on to underlying functions.}
}
\value{
A data table (tibble) with columns for each requested variable.
}
\description{
This function is a wrapper for all SEMRush reports functions included in this package. It takes a single argument ('report'), and passes
all other arguments to the associated '*_reports' functions.
}
\examples{
\dontrun{

report <-
  semrush_reports(
    report = "domain",
    type = "domain_organic",
    key = key,
    database = "us",
    domain = "cran.r-project.org"
  )

}

}
