% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_reports.R
\name{url_reports}
\alias{url_reports}
\title{Request SEMRush URL Reports}
\usage{
url_reports(
  type,
  key,
  url,
  database,
  display_limit = 5,
  display_offset,
  display_date,
  export_columns,
  return_url = FALSE
)
}
\arguments{
\item{type}{\emph{string}. REQUIRED. The report type to be generated. Available report types are 'url_organic' or 'url_adwords'.}

\item{key}{\emph{string}. REQUIRED. An identification key assigned to a user after subscribing to SEMrush. The key is available on the Profile page.}

\item{url}{\emph{string}. REQUIRED. The URL of a landing page you would like to investigate (example: https://cran.r-project.org/).}

\item{database}{\emph{string}. REQUIRED. A regional database. If this parameter is not specified (i.e., ""), your request will be sent to all regional databases.}

\item{display_limit}{\emph{integer}. The number of results returned to a request. If this parameter is set to 0, up to 10,000 lines will be returned. Default is 5 lines.}

\item{display_offset}{\emph{integer}. This parameter allows you to skip a specified number of results before sending a report. Please keep in mind that if you use the option display_offset, display_limit, the value should be increased by value display_offset.}

\item{display_date}{\emph{string}. Date in format "YYYYMM15". A date on which a report will be shown. You can roll back to the past or choose an actual date.}

\item{export_columns}{\emph{vector}. A vector of character strings specifying the variables to be included in the report, which vary according to the report type (see 'type' argument). If this parameter is not specified, default columns will be sent.}

\item{return_url}{\emph{logical}. If TRUE, prints the request URL used to generate the report. Default value is FALSE.}
}
\value{
A data table (tibble) with columns for each requested variable.
}
\description{
This function creates a request to the SEMRush SEO database for users with a subscription and sufficient API units.
See the SEMRush API website \href{https://www.semrush.com/api-analytics/}{(https://www.semrush.com/api-analytics/)} for additional information, including a list of codes for the regional databases and export variables.
}
\examples{
\dontrun{
#Enter your SEMRush account API key.
#'key <- ""

## Get 'url_organic' report
urlToInvestigate <- "https://cran.r-project.org/"
#request report
report <- url_reports(
    type = "url_organic",
    key = key,
    url = urlToInvestigate,
    database = "us",
    display_limit = 5,
    export_columns = c("Ph", "Po", "Pp", "Nq", "Cp")
)

print(report)
# A tibble: 5 x 5
 Keyword                    Position Previous.Position Search.Volume   CPC
 <chr>                         <int>             <int>         <int> <dbl>
1 cran                              1                 1          8100  1.42
2 r cran                            1                 1          5400  0   
3 r cran download                   1                 1           590  0   
4 cran r project org windows        1                 1           590  0   
5 cran r project                    1                 1           480  0
} 
}
